% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvar.R
\name{.plotdccfit}
\alias{.plotdccfit}
\title{In R my dccfit object is named fit1 i.e fit1 <- dccfit(spec=..., data=dat1, ...)
Adjusted code from rmgarch package on github that adds 5% VaR and removes 99% VaR from plot is found in this rmgadjust package
Changes only appear in the function .plot.dccfit.5
Changes also include labelling the axes, inserting a legend
Use the code rmgadjust:::.plot.dccfit.5(fit1, which=5) to get the new VaR plot with 1%, 5% and Returns
Since code does not appear in NAMESPACE but can be accessed internally, hence use :::
All changes are allowed by the author of rmgarch and rugarch packages, Alexios Ghalanos and is not responsible for any errors made.}
\usage{
.plotdccfit(x, which = "ask", series = c(1, 2), ...)
}
\arguments{
\item{x}{numeric variable}

\item{'which'}{is a charater string and you can check rmgarch package}

\item{'...'}{is a charater string and you can check rmgarch package}

\item{'series'}{is a charater string and you can check rmgarch package}
}
\description{
In R my dccfit object is named fit1 i.e fit1 <- dccfit(spec=..., data=dat1, ...)
Adjusted code from rmgarch package on github that adds 5% VaR and removes 99% VaR from plot is found in this rmgadjust package
Changes only appear in the function .plot.dccfit.5
Changes also include labelling the axes, inserting a legend
Use the code rmgadjust:::.plot.dccfit.5(fit1, which=5) to get the new VaR plot with 1%, 5% and Returns
Since code does not appear in NAMESPACE but can be accessed internally, hence use :::
All changes are allowed by the author of rmgarch and rugarch packages, Alexios Ghalanos and is not responsible for any errors made.
}
\keyword{internal}
